backups/requirements.txt
streamlit
PyMuPDF
Pillow
google-generativeai
pandas
openpyxl
requests
